+++ Yurchenko_Final/src/main/java/model/service/TestInfoService.java
@@ -7,7 +7,7 @@ public interface TestInfoService {
 
     List<TestInfo> findTestsInfoByParameter(String column, Object value);
 
-    List<TestInfo> findUserTestInfoForPagination(Long userId, int currentPage, int recordsPerPage);
+    List<TestInfo> findUserTestInfoForPagination(Integer userId, int currentPage, int recordsPerPage);
 
     List<TestInfo> findTestInfoForPagination(int currentPage, int recordsPerPage);
 }
